EC-12(unit tests)

+++ ElectronicCard/EC.BusinessLogic/Services/Implementation/RecordService.cs
@@ -1,4 +1,5 @@
-﻿using EC.BusinessLogic.Services.Interfaces;
+﻿using System;
+using EC.BusinessLogic.Services.Interfaces;
 using EC.DataAccess.Repositories.Interfaces;
 using EC.Entities.Entities;
 using System.Collections.Generic;
@@ -23,25 +24,41 @@ namespace EC.BusinessLogic.Services.Implementation
 
             var user = _userRepository.GetById(record.DoctorId);
 
-            if (user.Doctor.Patients.FirstOrDefault(p => p.Id == record.PatientId) == null)
+            if (user != null)
             {
-                _userRepository.AddPatientToDoctor(record.PatientId, user.Id);
+                if (user.Doctor.Patients.FirstOrDefault(p => p.Id == record.PatientId) == null)
+                {
+                    _userRepository.AddPatientToDoctor(record.PatientId, user.Id);
+                }
             }
         }
 
         public void UpdateRecord(Record record)
         {
-            _recordRepository.Update(record);
+            if (record != null)
+            {
+                _recordRepository.Update(record);
+            }
+            else
+            {
+                throw new NullReferenceException();
+            }
+            
         }
 
         public void DeleteRecord(int? id)
         {
+            if (id <= 0 || id == null)
+            {
+                return;
+            }
+
             _recordRepository.Delete(id);
         }
 
         public Record GetRecordById(int? id)
         {
-            return id == null ? null : _recordRepository.GetById(id);
+            return id == null || id <= 0 ? null : _recordRepository.GetById(id);
         }
 
         public IReadOnlyCollection<Record> GetAllRecords()
